fixes tests again from import weirdness
